Virus Bewegung

+++ Enemy.java
@@ -18,7 +18,7 @@ public abstract class Enemy extends Creature
      * @param speed die Geschwindigkeit des Gegners
      * @since 0.01 (15.05.2019)
      */
-    public Enemy(int x, int y, String name, SpriteSheet enemySprite, int health, int speed, Weapon weapon) 
+    public Enemy(int x, int y, String name, SpriteSheet enemySprite, int health, double speed, Weapon weapon) 
     {
         super(name, enemySprite, x, y, Entity.DEFAULT_WIDTH, Entity.DEFAULT_HEIGHT, health, speed, weapon);
     }
